requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.8
django-environ==0.4.5
djangorestframework==3.11.0
idna==2.10
isort==4.3.21
joblib==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.5
psycopg2-binary==2.8.5
pylint==2.5.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tqdm==4.48.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
